updated script

+++ practicas/p1_s4_mpi/p1_s4_cuda/src/MPI_execution_script.sh
@@ -1,23 +1,29 @@
 #!/bin/bash
 
 directory=$(pwd)
-# Ve al directorio donde se generan las matrices
-cd ..
-cd ..
-cd ..
-cd p1_s1/generadorMatrices
 
+if [ ! -e milaleatoria ] || [ ! -e milidentidad ] || [ ! -e diezmilaleatoria ] || [ ! -e diezmilidentidad ]; then
+	
+	# Ve al directorio donde se generan las matrices
+	cd ..
+	cd ..
+	cd ..
+	cd p1_s1/generadorMatrices
 
-# Genera matrices de 1000x1000 y 10000x10000
-./generadorMatrices 1000 1000 0 milaleatoria
-./generadorMatrices 1000 1000 1 milidentidad
+	# Genera matrices de 1000x1000 y 10000x10000
+	./generadorMatrices 1000 1000 0 milaleatoria
+	./generadorMatrices 1000 1000 1 milidentidad
 
-./generadorMatrices 10000 10000 0 diezmilaleatoria
-./generadorMatrices 10000 10000 1 diezmilidentidad
+	./generadorMatrices 10000 10000 0 diezmilaleatoria
+	./generadorMatrices 10000 10000 1 diezmilidentidad
 
+	# Mueve las matrices al directorio de ejecucion
+	mv milaleatoria milidentidad diezmilaleatoria diezmilidentidad $directory
 
-# Mueve las matrices al directorio de ejecucion
-mv milaleatoria milidentidad diezmilaleatoria diezmilidentidad $directory
+	#Sube a el directorio de ejecucion
+	cd $directory
+
+fi
 
 
 # Compila los archivos necesarios para la ejecucion
